Continue with unification of loading logic.
